+++ routes/src/lib.rs
@@ -4,3 +4,4 @@ pub mod faqs;
 pub mod food;
 pub mod printing;
 pub mod spaces;
+pub mod tools;
