+++ routes/src/tools.rs
@@ -1,0 +1,6 @@
+use axum::{routing::get, Router};
+use handlers::github::help_handler;
+
+pub fn router() -> Router {
+    Router::new().route("/help", get(help_handler))
+}
