+++ src/routes.rs
@@ -1,3 +1,2 @@
-pub mod tools;
 pub mod transit;
 pub mod webcams;
